Moved listener to it's own mod
